Split CUDA_INIT from cuda_utilities.
