+++ src/AFQMC/Memory/CUDA/cuda_utilities.cpp
@@ -29,8 +29,6 @@
 #include "cusparse.h"
 #include "cusolverDn.h"
 #include "curand.h"
-#include "mpi3/communicator.hpp"
-#include "mpi3/shared_communicator.hpp"
 
 #include "multi/array.hpp"
 #include "multi/array_ref.hpp"
@@ -59,54 +57,6 @@ namespace qmc_cuda {
                                          &afqmc_cusolverDn_handle,
                                          &afqmc_curand_generator  
                                         }; 
-  // need a cleanup routine
-  void CUDA_INIT(boost::mpi3::shared_communicator& node, unsigned long long int iseed)
-  {
-
-    if(afqmc_cuda_handles_init) return;
-    afqmc_cuda_handles_init=true;
-
-    int num_devices=0;
-    cudaGetDeviceCount(&num_devices);
-    qmcplusplus::app_log()<<" Running in node with " <<num_devices <<" GPUs. \n";
-    if(num_devices < node.size()) {
-      qmcplusplus::app_error()<<"Error: # GPU < # tasks in node. " <<std::endl;
-      qmcplusplus::app_error()<<"# GPU: " <<num_devices <<std::endl;
-      qmcplusplus::app_error()<<"# tasks: " <<node.size() <<std::endl;
-      APP_ABORT("");
-    } else if(num_devices > node.size()) {
-      qmcplusplus::app_log()<<"WARNING: Unused devices !!!!!!!!!!!!!! \n"
-                                <<"         # tasks: " <<node.size() <<"\n"
-                                <<"         num_devices: " <<num_devices <<std::endl;
-    }
-
-    cuda_check(cudaSetDevice(node.rank()),"cudaSetDevice()");
-
-    cublas_check(cublasCreate (& afqmc_cublas_handle ), "cublasCreate");
-//    cublas_check(cublasXtCreate (& afqmc_cublasXt_handle ), "cublasXtCreate");
-    int devID[8] {0,1,2,3,4,5,6,7};
-//    cublas_check(cublasXtDeviceSelect(afqmc_cublasXt_handle, 1, devID), "cublasXtDeviceSelect");
-//    cublas_check(cublasXtSetPinningMemMode(afqmc_cublasXt_handle, CUBLASXT_PINNING_ENABLED), 
-//                                            "cublasXtSetPinningMemMode");
-    cusolver_check(cusolverDnCreate (& afqmc_cusolverDn_handle ), "cusolverDnCreate");
-    //curand_check(curandCreateGenerator(&afqmc_curand_generator, CURAND_RNG_PSEUDO_DEFAULT),
-    curand_check(curandCreateGenerator(&afqmc_curand_generator, CURAND_RNG_PSEUDO_MT19937),
-                                            "curandCreateGenerator");
-    curand_check(curandSetPseudoRandomGeneratorSeed(afqmc_curand_generator,iseed),
-                                            "curandSetPseudoRandomGeneratorSeed");
-
-    cusparse_check(cusparseCreate (& afqmc_cusparse_handle ), "cusparseCreate");
-    cusparse_check(cusparseCreateMatDescr(&afqmc_cusparse_matrix_descr), 
-            "cusparseCreateMatDescr: Matrix descriptor initialization failed"); 
-    cusparseSetMatType(afqmc_cusparse_matrix_descr,CUSPARSE_MATRIX_TYPE_GENERAL);
-    cusparseSetMatIndexBase(afqmc_cusparse_matrix_descr,CUSPARSE_INDEX_BASE_ZERO); 
-
-    cusparse_buffer = new boost::multi::array<std::complex<double>,1,
-                                 qmc_cuda::cuda_gpu_allocator<std::complex<double>>>(
-                                 (typename boost::multi::layout_t<1u>::extensions_type{1},
-                                 qmc_cuda::cuda_gpu_allocator<std::complex<double>>{}));
-
-  }
 
   void cuda_check(cudaError_t sucess, std::string message)
   {
