+++ src/AFQMC/Memory/custom_pointers.hpp
@@ -18,6 +18,7 @@
 #include "AFQMC/Memory/raw_pointers.hpp" 
 #if defined(ENABLE_CUDA)
 #include "AFQMC/Memory/CUDA/cuda_utilities.h"
+#include "AFQMC/Memory/CUDA/cuda_init.h"
 #include "AFQMC/Memory/CUDA/cuda_gpu_pointer.hpp"
 #endif
